backend/requirements.txt
# --- 核心框架 ---
Django==4.2.9            # LTS 长期支持版
djangorestframework==3.14.0
django-cors-headers==4.3.1

# --- 数据库与缓存 ---
# MySQL数据库驱动（Windows推荐使用pymysql，Linux可以使用mysqlclient）
pymysql==1.1.0
# mysqlclient==2.2.1  # Linux/Mac推荐，Windows安装可能较困难
django-redis==5.4.0

# --- AI 与 工具 ---
google-generativeai      # Gemini API SDK
openai==1.12.0           # (可选) OpenAI SDK
python-dotenv==1.0.1     # 环境变量管理
Pillow==10.2.0           # 图片处理
requests==2.31.0         # 网络请求
pandas==2.1.4            # Excel文件处理
openpyxl==3.1.2          # Excel文件读取支持